Rework the section field structure
They're all the same except for the value type, which most of the time is a
string. Extracting the name out of each variant becomes unwieldy as we would
need to match.

This way, where we have more dynamic types, we should match better with the way
the data is laid out.